% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antismash_parseR _function.R
\name{antismash_parseR}
\alias{antismash_parseR}
\title{Parse antismash output.}
\usage{
antismash_parseR(
  dir_name,
  accession = NA,
  output_dir = NA,
  output_product = TRUE,
  output_orfs = TRUE,
  output_gbks = FALSE,
  include_seqs = FALSE
)
}
\arguments{
\item{dir_name}{required dir path; antismash output directory}

\item{accession}{optional string; accession or unique ID for the output, for naming the outputs}

\item{output_dir}{optional dir path; directory for saving output files; defaults to directory specified in dir_name}

\item{output_product}{TRUE/FALSE; output products for each BGC region. defaults TRUE.}

\item{output_orfs}{TRUE/FALSE; output all orfs in each BGC region as specified in the regions.js file. Sometimes antismash changes locus tags for some genes. defaults TRUE.}

\item{output_gbks}{TRUE/FALSE; output all orfs in the GBK files for each BGC region. Contains extra border genes. Also contains unmodified locus tags for each gene. defaults FALSE.}

\item{include_seqs}{TRUE/FALSE; include protein translation in the output_orfs csv for each orf in the BGC region. defaults FALSE.}
}
\value{
csv files in the output directory that contain gene IDs in each BGC region or backbone enzyme types.

Three possible outputs, depending on function arguments above:

antismash_parseR_all_orfs.csv; c("Accession", "cluster", "chrm", "region", "locus_tag", "gene", "type", "bb_type", "start", "end", "strand", "translation")

antismash_parseR_all_products.csv; c("Accession", "cluster", "chrm", "region", "product", "product_category", "backbone_locus_tag", "backbone_gene", "bb_type")

antismash_parseR_all_gbk_genes.csv; c("Accession", "chrm", "region", "locus_tag", "gene", "start_end_strand", "excluded_from_cluster", "cluster")
}
\description{
This function parses antismash output directories to find lists of genes, products, and backbone enzymes and imports them to R.
This function assumes you are running it on the antismash results directory.
Specifically, it parses the regions.js file and all of the region .gbk files.
These files must all be present and not compressed in the output directory.
Other files are not necessary.
}
