% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_bb_type_function.R
\name{find_bb_type}
\alias{find_bb_type}
\title{Find backbone enzyme types.}
\usage{
find_bb_type(index, all_orfs)
}
\arguments{
\item{index}{gene index in the table}
}
\value{
vector list flattened by ;
}
\description{
This function is used within antismash_parseR to find backbone enzyme types.
for specific genes, parse the description column and extract the enzyme type.
split by "\if{html}{\out{<br>}}" and keep biosynthetic regions.
}
