% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_gbk_function.R
\name{import_gbk}
\alias{import_gbk}
\title{Import antismash gbk files.}
\usage{
import_gbk(gbk_file)
}
\arguments{
\item{gbk_file}{path to gbk_file}
}
\value{
genes and associated info extracted from the gbk
}
\description{
This function is used within antismash_parseR to import antismash gbk files and extract ORF info.
}
