% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_busco_function.R
\name{parse_busco}
\alias{parse_busco}
\title{Parse BUSCO output.}
\usage{
parse_busco(file_name, accession = NA)
}
\arguments{
\item{file_name}{required file path; short_summary.txt output file from BUSCO}

\item{accession}{optional string; accession or unique ID for the output}
}
\value{
A dataframe with the busco results
}
\description{
This function parses individual busco summary tables and imports them to R.
}
