% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_quast_function.R
\name{parse_quast}
\alias{parse_quast}
\title{Parse Quast output.}
\usage{
parse_quast(file_name, accession = NA)
}
\arguments{
\item{file_name}{required file path; report.tsv output file from quast}

\item{accession}{optional string; accession or unique ID for the output}
}
\value{
A dataframe with the quast statistics
}
\description{
This function parses individual quast output tables and imports them to R.
}
